+++ BitirmeProjesiBackend/Application/Dtos/EventDto.cs
@@ -14,7 +14,7 @@
         public int Point { get; set; }
         public int Quota { get; set; }
 
-        public int Longitude { get; set; }
-        public int Latitude { get; set; }
+        public float Longitude { get; set; }
+        public float Latitude { get; set; }
     }
 }
